Use models.py for main classes.
